split object.rs into several files
